+++ package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "files": [
-    "src/*.js",
+    "src",
     "dist"
   ],
   "scripts": {
@@ -92,7 +92,7 @@
   },
   "jest": {
     "roots": [
-      "./src"
+      "./tests"
     ],
     "setupFiles": [
       "raf/polyfill"
