+++ tests/ReactDOMMock.js
