+++ tests/web.test.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as plug from 'react-powerplug';
 import TestRenderer from 'react-test-renderer';
 import './ReactDOMMock';
-import { HeadProvider, Title, Style, Meta, Link } from '../';
+import { HeadProvider, Title, Style, Meta, Link } from '../src';
 
 describe('head tag during client', () => {
   it('renders into document.head portal', () => {
